requirements.txt
flask
pandas
numpy
scikit-learn
xgboost
joblib
tqdm
pyyaml
requests
